Move analiza-zwolnien column definitions to a new file
