+++ src/app/analiza-zwolnien/analiza-zwolnien.component.html
@@ -3,7 +3,7 @@
   class="ag-theme-quartz"
   [rowData]="companies()"
   [rowHeight]="rowHeight"
-  [columnDefs]="columnDefs"
+  [columnDefs]="colDefs"
   [defaultColDef]="defaultColDef"
   (cellValueChanged)="onCellValueChanged($event)"
 >
